requirements.txt
Django==5.2
django-ninja==1.4.3
psycopg2-binary==2.9.10
django-environ==0.12.0
aiogram==3.20.0.post0
aiogram_dialog==2.3.1